+++ app/src/main/kotlin/com/maltaisn/notes/ui/edit/adapter/EditListViewHolder.kt
@@ -16,10 +16,12 @@
 
 package com.maltaisn.notes.ui.edit.adapter
 
+import android.graphics.Paint
 import android.text.Editable
 import android.text.InputFilter
 import android.text.Spannable
 import android.view.KeyEvent
+import android.widget.EditText
 import androidx.core.text.clearSpans
 import androidx.core.view.isInvisible
 import androidx.core.widget.doOnTextChanged
@@ -107,6 +109,8 @@ class EditItemViewHolder(val binding: ItemEditItemBinding, callback: EditAdapter
     init {
         itemCheck.setOnCheckedChangeListener { _, isChecked ->
             item.checked = isChecked
+            itemEdt.strikethroughText = isChecked && callback.strikethroughCheckedItems
+            itemEdt.isActivated = !isChecked // Controls text color selector.
         }
 
         itemEdt.filters = arrayOf(clearSpansInputFilter)
@@ -161,6 +165,7 @@ class EditItemViewHolder(val binding: ItemEditItemBinding, callback: EditAdapter
 
         itemCheck.isChecked = item.checked
         itemCheck.isEnabled = item.editable
+        itemEdt.isActivated = !item.checked
     }
 
     override fun setFocus(pos: Int) {
@@ -203,3 +208,13 @@ private val clearSpansInputFilter = InputFilter { source, start, end, _, _, _ ->
     }
     filtered
 }
+
+private var EditText.strikethroughText: Boolean
+    get() = (this.paintFlags and Paint.STRIKE_THRU_TEXT_FLAG) != 0
+    set(value) {
+        this.paintFlags = if (value) {
+            this.paintFlags or Paint.STRIKE_THRU_TEXT_FLAG
+        } else {
+            this.paintFlags and Paint.STRIKE_THRU_TEXT_FLAG.inv()
+        }
+    }